pyproject.toml
[tool.poetry]
name = "kingdom-rl"
version = "0.1.0"
description = ""
authors = ["Arina <arina.iusupova.ai@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
pydirectinput-rgx = "^2.1.2"
pywin32 = "^308"
pre-commit = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
